Finished automating the data processing pipeline
