--- data_processing/data_processing/utils.py
@@ -1,19 +1,0 @@
-# TODO: Delete this
-
-from dataclasses import dataclass
-from typing import Tuple
-
-@dataclass
-class Point:
-    x: int
-    y: int
-
-@dataclass
-class Region:
-    top_left: Point
-    bottom_right: Point
-
-@dataclass
-class Resolution:
-    height: int
-    width: int
